pyproject.toml
[tool.poetry]
name = "brics-tools"
version = "0.0.0"
description = "Various modules and scripts to search within a Biomedical Research Informatics Computing System (BRICS) instance."
authors = ["Kevin Armengol"]
readme = "README.md"
packages = [{include = "brics_tools"}]

[tool.poetry.dependencies]
hydra-core = ">=1.1.0"
omegaconf = "2.1.2"
openpyxl = ">=3.0.10"
pandas = ">=1.5.2"
pathlib = "^1.0.1"
pinecone-client = "^2.2.1"
python = ">=3.8.1,!=3.9.7, <3.12"
python-dotenv = "0.21.1"
requests = ">=2.28.1"
requests-html = ">=0.10.0"
sentence-transformers = "^2.2.2"
ipykernel = "^6.23.1"
joblib = "^1.2.0"
matplotlib = "^3.7.2"
chromadb = ">=0.4.4"
llama-index = ">=0.7.19"
llama-cpp-python = "^0.1.77"
replicate = "^0.11.0"
span-marker = "^1.2.4"
nest-asyncio = "^1.5.7"
arize-phoenix = "^0.0.32"
streamlit = "^1.26.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
bs4 = "^0.0.1"
deptry = "^0.8.0"
flake8 = "^6.0.0"
isort = "^5.11.3"
jupyter = "^1.0.0"
jupyterlab = ">=3.6.1"
monkeytype = "^23.3.0"
pre-commit = "^3.0.4"
radon = "^5.1.0"
vulture = "^2.7"
wheel = "^0.38.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
